feature/#99 feat: add useDecks
